chaingen: Add revocation generation infrastructure.
This adds infrastructure to track missed votes and automatically
generate revocations for them along with a new assertion helper named
AssertTipNumRevocations which, as it name suggests, allows the caller to
assert the header commits to the specified number of revocations.

This will be useful for porting the rest of the block validation tests
to the fullblocktests framework.