mining: Correct initial template generation.
This corrects the code to create a new template when first starting up
with aggressive mining enable before votes on the most recent block are
known.

First, the code is expecting the current best block, not its parent, and
second, the coinbase script should match what is created in the normal
path as well.