fullblocktests: Add expired stake tx test.
This adds a test to ensure blocks with an expired transaction in the
stake tree is rejected as expected.

While here, it also updates the existing test for an expired transaction
in the regular tree to be the exact block height which is including it
to help detect any off by one errors.