stake: add ExpiredByBlock to stake.Node
This adds the (*Node).ExpiredByBlock to the stake package. MissedByBlock
includes both missed (a winner with no vote) and expired. ExpiredByBlock
gives only tickets that expired.

Note that these functions check the undo data of a stake.Node so they
results reflect the latest block to be connected and written to the
backing database.

When a ticket initially expires, it is tagged as Missed and Expired,
then put into the missedTickets bucket.

Later when and if it is revoked, it is extracted from the missed tickets
bucket, retaining both the Missed and Expired flags, marked Revoked, and
put into the revokedTickets bucket.

When checking the undo data with ExpiredByBlock, it is thus necessary to
check (Expired && ! Revoked), otherwise the same ticket may be reported
as expired twice by different blocks.

The exact same issue with semantic affects MissedByBlock, just for both
expires and real misses.

A similar problem does NOT affect SpendByBlock as this applies only to
votes, which are just removed from the liveTickets bucket, but not moved
from bucket to bucket.

This includes tests for ExpiredByBlock, MissedByBlock, and SpentByBlock.
These are done in a simnet chain with 1001 blocks to observe expiry.