blockchain: Panic on fatal assertions.
This modifies the code to panic under unrecoverable conditions which
will very likely lead to worse things happening such as writing
incorrect information to the database versus return an assertion error.

We generally prefer to avoid using panics in packages since it largely
takes control away from the caller (although they can use recover to
catch them) and they are often misused as a substitute for proper error
handling.  However, the particular cases this changes really need to be
panics because they apply to algorithmic invariants that if violated can
lead to consensus failure, corrupted state, and other disastrous
outcomes.

Note that there are still some other cases in the package that return
assert errors, but they either really should be returning a different
error as they are not unrecoverable situations or they're in code that
will be removed.