blockchain: Improve spend journal tests.
This cleans up and corrects the spend journal tests which deal with a
live blockchain.

First, it updates the params used in the test to the legacy values so
the blocks actually connect to the main chain instead of being added as
orphans.

Second, it checks the return value from ProcessBlock for whether or not
the block is on the main chain and fails if it is not to ensure it
actually is on the main chain.

Finally, it moves the core logic which checks the number of spend
outputs matches the serialized spend journal directly into the test
instead of calling a separate DoStxoTest function on the blockchain
instance for that purpose and removes that now unused function.