server: Minor function definition order cleanup.
This moves the function definition for enforceNodeCFFlag so it are more
consistent with the preferred order used throughout the codebase.  In
particular, the functions are defined before they're first used and
generally as close as possible to the first use when they're defined in
the same file.

While moving, it also reformats the code to match the same spacing as
the other CF functions for consistency.

There are no functional changes.