multi: revert TicketChange addition to PurchaseTicketCmd
This reverts the split_tx/ticket_change addition to PurchaseTickeCmd. From a discussion had with raedah and jrick on purchasing tickets we concluded that change created by direct ticket purchases will adhere to coinbase maturity before becoming available for spending. We will not be circumventing this by not setting the expiry for sstxchange.

The only ideal scenario currently for purchasing tickets directly is with inputs equal to the stake difficulty plus the fee.