txscript: Make PeekInt consistent with PopInt.
This modifies the PeekInt function of the stack to accept a maximum
script number length to mirror PopInt for consistency.  It also updates
the two callers CLTV and CSV) which were manually performing the same
task with 5 bytes due to PeekInt enforcing 4-byte script nums to use the
modified version accordingly.

It also adds some stack tests for 5-byte encodings on both PopInt and
PeekInt.