blockchain: Implement new chain view.
This implements a new type in the blockchain package that takes
advantage of the fact that all block nodes are now in memory to provide
a flat view of a specific chain of blocks (a specific branch of the
overall block tree) from a given tip all the way back to the genesis
block along with several convenience functions such as efficiently
comparing two views, quickly finding the fork point (if any) between two
views, and O(1) lookup of the node at a specific height.

The view is not currently used, but the intent is that the code will be
refactored to make use of these views to simplify and optimize several
areas such as best chain selection and reorg logic and finding successor
nodes.  They will also greatly simplify the process of disconnecting the
download logic from the connection logic.

Since the ultimate intent is for there to be a long-lived chain view
instance for the current best chain, this also implements efficient
handling of setting the tip to new values including domain-specific
capacity increase handling which chooses the desired increase amount
more intelligently than the default algorithm, which would way
overshoot, as well as adds some additional space when the view is
initialized.

A comprehensive suite of tests is provided to ensure the chain views
behave correctly.