blockchain: Remove unnecessary genesis block check.
This modifies the checkConnectBlock function to remove the check which
tests if the block is the genesis block and returns an error since it is
no longer necessary due to other changes in the code since the function
was originally written.

In particular, the genesis block is valid by definition as the root of
the chain, so it does not make sense to try to connect it to a chain
that doesn't exist yet.