blockchain: Make block index flushable.
This modifies the block index to support tracking modified nodes and
flushing them to the database.  This ensures that validation states set
on arbitrary nodes are persisted to the database and helps pave the way
towards supporting decoupling the chain processing logic from the
download logic.

In the immediate term, as of this commit, side chains which encounter an
invalid block will no longer potentially have to be revalidated to
discover they are invalid after restart.

A few comments are also improved while here.